requirements.txt
graphviz
networkx 
numpy
scipy
pandas
torch
hydra-core
wandb
yaml
einops